pyproject.toml
[project]
name = "ai-hedge-fund"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.16.0",
    "litellm>=1.78.7",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
